+++ NEWS.md
@@ -1,3 +1,17 @@
+# tidyREDCap 1.1.0.9000
+
+## New features
+
+* Add `make_yes_no()` function to convert "checked" or "yes"-like answers to 
+  "Yes" and other answers to "No or Unknown".
+* Add `make_yes_no_unknown()` function to convert "checked" or "yes"-like 
+  answers to "Yes"", unchecked or "no"-like answers to "No" and other answers to
+  "Unknown".
+  
+## Fixes/Changes
+* `make_choose_all_table()` now works with api or manual/point-and-click 
+  exports.
+
 # tidyREDCap 1.1.0  (CRAN release)
 
 ## New features